+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 **Date:** 2 April 2026  
 **Description:**  
 FEATURE UPDATE: Added r2livecode.py - Navigation Algorithm for Final Maze
+
+## *v1.2.0*  
+**Date:** 7 April 2026  
+**Description**  
+FEATURE UPDATE: Added r2dockingtest.py for robot docking [WORK IN PROGRESS]  
+REFINEMENTS: Parameter Updates to r2livecode